refactor: CookieManager 패키지 이동
